move chat html css to flyte.ai.chat
