+++ src/flyte/ai/agents/__init__.py
@@ -1,12 +1,10 @@
 """flyte.ai.agents — Agent abstractions and chat UI for Flyte apps."""
 
-from ._css import DEFAULT_CSS
 from .chat_ui import AgentChatAppEnvironment, CustomTheme
 from .codemode import CodeModeAgent
 from .protocol import Agent, AgentResult
 
 __all__ = [
-    "DEFAULT_CSS",
     "Agent",
     "AgentChatAppEnvironment",
     "AgentResult",
