+++ src/flyte/ai/chat/__init__.py
@@ -1,0 +1,11 @@
+"""flyte.ai.chat — HTML/CSS assets for the Agent Chat UI."""
+
+from ._css import CUSTOM_THEME_CSS_TEMPLATE, DEFAULT_CSS
+from ._html import CHAT_HTML_TEMPLATE, build_chat_html
+
+__all__ = [
+    "CHAT_HTML_TEMPLATE",
+    "CUSTOM_THEME_CSS_TEMPLATE",
+    "DEFAULT_CSS",
+    "build_chat_html",
+]
