+++ src/flyte/ai/chat/_html.py
